Feature: The Combinational Analysis module supports apostrophes (as in "x'" for NOT x).
Feature: For the I/O Library's LED Matrix, the Input Format attribute can be changed after the matrix is added into the circuit.

Bug fix: Repaired so that the documented "-clearprops" command-line option actually works. (It was programmed to accept "-clearprefs" but it was described as "-clearprops" in the "-help" text and in the documentation.)

Design change: The Combinational Analysis module's expression pane uses a larger font size than previously.